pyproject.toml
[project]
name = "ai-employee-system"
version = "0.2.0"
description = "A Personal AI Employee system implementing autonomous FTE capabilities - Gold Tier"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Existing dependencies
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "linkedin-api>=2.3.1",
    "playwright>=1.58.0",
    "watchdog>=4.0.0",
    
    # Gold Tier: HTTP and API clients
    "requests>=2.32.0",
    
    # Gold Tier: Social Media APIs
    "facebook-business>=19.0.0",
    "tweepy>=4.14.0",
    
    # Gold Tier: Retry logic
    "retrying>=1.3.0",
    
    # Gold Tier: Structured logging
    "structlog>=24.0.0",
    
    # Gold Tier: OAuth for social media
    "requests-oauthlib>=1.3.1",
    
    # Gold Tier: Image processing for Instagram
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "playwright>=1.58.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
